git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@4294 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/arm/include/chronometer.h
@@ -28,14 +28,7 @@ public:
     void lap() { if(_start != 0) _stop = tsc.time_stamp(); }
     void stop() { lap(); }
 
-    Microsecond read() 
-    { 
-        // The parenthesis may cause a precision loss, but enables
-        // the compiler to perform the division.
-        // This code causes an exception in eMote3 without the
-        // parenthesis
-        return ticks() * (1000000ULL / frequency()); 
-    }
+    Microsecond read() { return ticks() * 1000000 / frequency(); }
 
 private:
     Time_Stamp ticks() {
@@ -71,11 +64,8 @@ public:
     void lap() { if(_start != 0) _stop = Alarm::_elapsed; }
     void stop() { lap(); }
 
-    // The parenthesis may cause a precision loss, but enables
-    // the compiler to perform the division.
-    // This code causes an exception in eMote3 without the
-    // parenthesis
-    Microsecond read() { return ticks() * (1000000ULL / frequency()); }
+    // The cast provides resolution for intermediate calculations
+    Microsecond read() { return static_cast<unsigned long long>(ticks()) * 1000000 / frequency(); }
 
 private:
     Time_Stamp ticks() {
